ml-model/vehicle-recognition/requirements.txt
torch
torchvision
boto3
pillow